+++ OUDS/Core/Themes/Orange/Tests/Values/SemanticTokens/ThemeOverrideOfColorModeSemanticTokensTests.swift
@@ -36,6 +36,50 @@ struct ThemeOverrideOfColorModeSemanticTokensTests {
 
     // MARK: - Tests
 
+    // MARK: - On Background
+
+    @Test func inheritedThemeCanOverrideModeOnBgPrimaryLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgPrimaryLight != abstractTheme.colorModes.modeOnBgPrimaryLight)
+        #expect(inheritedTheme.colorModes.modeOnBgPrimaryLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgPrimaryDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgPrimaryDark != abstractTheme.colorModes.modeOnBgPrimaryDark)
+        #expect(inheritedTheme.colorModes.modeOnBgPrimaryDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgSecondaryLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgSecondaryLight != abstractTheme.colorModes.modeOnBgSecondaryLight)
+        #expect(inheritedTheme.colorModes.modeOnBgSecondaryLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgSecondaryDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgSecondaryDark != abstractTheme.colorModes.modeOnBgSecondaryDark)
+        #expect(inheritedTheme.colorModes.modeOnBgSecondaryDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgTertiaryLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgTertiaryLight != abstractTheme.colorModes.modeOnBgTertiaryLight)
+        #expect(inheritedTheme.colorModes.modeOnBgTertiaryLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgTertiayDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgTertiaryDark != abstractTheme.colorModes.modeOnBgTertiaryDark)
+        #expect(inheritedTheme.colorModes.modeOnBgTertiaryDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgEmphasizedLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgEmphasizedLight != abstractTheme.colorModes.modeOnBgEmphasizedLight)
+        #expect(inheritedTheme.colorModes.modeOnBgEmphasizedLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnBgEmphasizedDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnBgEmphasizedDark != abstractTheme.colorModes.modeOnBgEmphasizedDark)
+        #expect(inheritedTheme.colorModes.modeOnBgEmphasizedDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    // MARK: - On Brand
+
     @Test func inheritedThemeCanOverrideModeOnBrandPrimaryLight() throws {
         #expect(inheritedTheme.colorModes.modeOnBrandPrimaryLight != abstractTheme.colorModes.modeOnBrandPrimaryLight)
         #expect(inheritedTheme.colorModes.modeOnBrandPrimaryLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
@@ -66,6 +110,8 @@ struct ThemeOverrideOfColorModeSemanticTokensTests {
         #expect(inheritedTheme.colorModes.modeOnBrandTertiaryDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
     }
 
+    // MARK: - On Status Accent
+
     @Test func inheritedThemeCanOverrideModeOnStatusAccentEmphasizedLight() throws {
         #expect(inheritedTheme.colorModes.modeOnStatusAccentEmphasizedLight != abstractTheme.colorModes.modeOnStatusAccentEmphasizedLight)
         #expect(inheritedTheme.colorModes.modeOnStatusAccentEmphasizedLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
@@ -85,6 +131,8 @@ struct ThemeOverrideOfColorModeSemanticTokensTests {
         #expect(inheritedTheme.colorModes.modeOnStatusAccentMutedDark != abstractTheme.colorModes.modeOnStatusAccentMutedDark)
         #expect(inheritedTheme.colorModes.modeOnStatusAccentMutedDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
     }
+
+    // MARK: - On Status
 
     @Test func inheritedThemeCanOverrideModeOnStatusInfoEmphasizedLight() throws {
         #expect(inheritedTheme.colorModes.modeOnStatusInfoEmphasizedLight != abstractTheme.colorModes.modeOnStatusInfoEmphasizedLight)
@@ -184,6 +232,40 @@ struct ThemeOverrideOfColorModeSemanticTokensTests {
     @Test func inheritedThemeCanOverrideModeOnStatusWarningMutedDark() throws {
         #expect(inheritedTheme.colorModes.modeOnStatusWarningMutedDark != abstractTheme.colorModes.modeOnStatusWarningMutedDark)
         #expect(inheritedTheme.colorModes.modeOnStatusWarningMutedDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    // MARK: - On Overlay
+
+    @Test func inheritedThemeCanOverrideModeOnOverlayDefaultLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnOverlayDefaultLight != abstractTheme.colorModes.modeOnOverlayDefaultLight)
+        #expect(inheritedTheme.colorModes.modeOnOverlayDefaultLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnOverlayDefaultDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnOverlayDefaultDark != abstractTheme.colorModes.modeOnOverlayDefaultDark)
+        #expect(inheritedTheme.colorModes.modeOnOverlayDefaultDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnOverlayEmphasizedLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnOverlayEmphasizedLight != abstractTheme.colorModes.modeOnOverlayEmphasizedLight)
+        #expect(inheritedTheme.colorModes.modeOnOverlayEmphasizedLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnOverlayEmphasizedDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnOverlayEmphasizedDark != abstractTheme.colorModes.modeOnOverlayEmphasizedDark)
+        #expect(inheritedTheme.colorModes.modeOnOverlayEmphasizedDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    // MARK: - On Modal
+
+    @Test func inheritedThemeCanOverrideModeOnModalLight() throws {
+        #expect(inheritedTheme.colorModes.modeOnModalLight != abstractTheme.colorModes.modeOnModalLight)
+        #expect(inheritedTheme.colorModes.modeOnModalLight == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
+    }
+
+    @Test func inheritedThemeCanOverrideModeOnModalDark() throws {
+        #expect(inheritedTheme.colorModes.modeOnModalDark != abstractTheme.colorModes.modeOnModalDark)
+        #expect(inheritedTheme.colorModes.modeOnModalDark == MockThemeColorModeSemanticTokensProvider.mockThemeColorModeSemanticToken)
     }
 }