+++ OUDS/Foundations/Sources/Extensions/String+SwiftUI.swift
@@ -21,9 +21,11 @@ extension String {
     /// `String` extension to get a `Color` from `self` supposed to be an hexadecimal string representation.
     /// We assume the value of `self` is a valid hexadecimal string.
     public var color: Color! {
+        #if DEBUG
         if self == "ouds-forbidden-color-value" {
             OL.error("Trying to parse forbidden color as SwiftUI Color, it will crash")
         }
+        #endif
         return Color(hexadecimalCode: self)
     }